Impl Tokens on TakeWhile, add Take, add single parse method
